pyproject.toml
[tool.poetry]
name = "diplom-ob1"
version = "0.1.0"
description = ""
authors = ["Denis Antropov <de94onisus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-filter = "^24.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.4.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"


[tool.poetry.group.linters.dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"


[tool.poetry.group.tests.dependencies]
coverage = "^7.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs         # исключить несколько общих каталогов в
    | \.git          # корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # также отдельно исключить файл с именем foo.py в
                     # корне проекта
)
'''

[tool.isort]
line_length = 119
